+++ .gitignore
@@ -1,0 +1,4 @@
+/.classpath
+/.project
+/target/
+/.settings/
