--- src/GameApp.java
@@ -1,14 +1,0 @@
-
-import javax.sound.sampled.LineUnavailableException;
-import javax.sound.sampled.UnsupportedAudioFileException;
-import java.io.IOException;
-
-public class GameApp {
-
-    public static void main(String[] args) throws UnsupportedAudioFileException, IOException, LineUnavailableException, InterruptedException {
-        Window win=new Window();
-       // Sound s=new Sound();
-        //win.playSound("C:\\java\\hang.wav");
-    }
-
-}
