+++ src/main/java/hu/peteri/pacman/Field.java
@@ -1,3 +1,4 @@
+package hu.peteri.pacman;
 // one piece of field, represents the unit block of the game board.
 //knows its own type(wall,empty,cherry...)
 //knows the next possible correct ways from itself
