+++ src/main/resources/pac_right.png
