+++ src/main/resources/pacman.png
